@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 class DrawerUserController extends StatefulWidget {
   const DrawerUserController({
-    Key key,
+    Key? key,
     this.drawerWidth = 250,
     this.onDrawerCall,
     this.screenView,
@@ -15,21 +15,21 @@ class DrawerUserController extends StatefulWidget {
   }) : super(key: key);
 
   final double drawerWidth;
-  final Function(DrawerIndex) onDrawerCall;
-  final Widget screenView;
-  final Function(bool) drawerIsOpen;
-  final AnimatedIconData animatedIconData;
-  final Widget menuView;
-  final DrawerIndex screenIndex;
+  final Function(DrawerIndex)? onDrawerCall;
+  final Widget? screenView;
+  final Function(bool)? drawerIsOpen;
+  final AnimatedIconData? animatedIconData;
+  final Widget? menuView;
+  final DrawerIndex? screenIndex;
 
   @override
   _DrawerUserControllerState createState() => _DrawerUserControllerState();
 }
 
 class _DrawerUserControllerState extends State<DrawerUserController> with TickerProviderStateMixin {
-  ScrollController scrollController;
-  AnimationController iconAnimationController;
-  AnimationController animationController;
+  ScrollController? scrollController;
+  AnimationController? iconAnimationController;
+  AnimationController? animationController;
 
   double scrolloffset = 0.0;
 
@@ -37,41 +37,41 @@ class _DrawerUserControllerState extends State<DrawerUserController> with Ticker
   void initState() {
     animationController = AnimationController(duration: const Duration(milliseconds: 2000), vsync: this);
     iconAnimationController = AnimationController(vsync: this, duration: const Duration(milliseconds: 0));
-    iconAnimationController..animateTo(1.0, duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
+    iconAnimationController?..animateTo(1.0, duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
     scrollController = ScrollController(initialScrollOffset: widget.drawerWidth);
-    scrollController
+    scrollController!
       ..addListener(() {
-        if (scrollController.offset <= 0) {
+        if (scrollController!.offset <= 0) {
           if (scrolloffset != 1.0) {
             setState(() {
               scrolloffset = 1.0;
               try {
-                widget.drawerIsOpen(true);
+                widget.drawerIsOpen!(true);
               } catch (_) {}
             });
           }
-          iconAnimationController.animateTo(0.0, duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
-        } else if (scrollController.offset > 0 && scrollController.offset < widget.drawerWidth.floor()) {
-          iconAnimationController.animateTo((scrollController.offset * 100 / (widget.drawerWidth)) / 100,
+          iconAnimationController?.animateTo(0.0, duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
+        } else if (scrollController!.offset > 0 && scrollController!.offset < widget.drawerWidth.floor()) {
+          iconAnimationController?.animateTo((scrollController!.offset * 100 / (widget.drawerWidth)) / 100,
               duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
         } else {
           if (scrolloffset != 0.0) {
             setState(() {
               scrolloffset = 0.0;
               try {
-                widget.drawerIsOpen(false);
+                widget.drawerIsOpen!(false);
               } catch (_) {}
             });
           }
-          iconAnimationController.animateTo(1.0, duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
+          iconAnimationController?.animateTo(1.0, duration: const Duration(milliseconds: 0), curve: Curves.fastOutSlowIn);
         }
       });
-    WidgetsBinding.instance.addPostFrameCallback((_) => getInitState());
+    WidgetsBinding.instance?.addPostFrameCallback((_) => getInitState());
     super.initState();
   }
 
   Future<bool> getInitState() async {
-    scrollController.jumpTo(
+    scrollController?.jumpTo(
       widget.drawerWidth,
     );
     return true;
@@ -96,18 +96,18 @@ class _DrawerUserControllerState extends State<DrawerUserController> with Ticker
                 //we divided first drawer Width with HomeDrawer and second full-screen Width with all home screen, we called screen View
                 height: MediaQuery.of(context).size.height,
                 child: AnimatedBuilder(
-                  animation: iconAnimationController,
-                  builder: (BuildContext context, Widget child) {
+                  animation: iconAnimationController!,
+                  builder: (BuildContext context, Widget? child) {
                     return Transform(
                       //transform we use for the stable drawer  we, not need to move with scroll view
-                      transform: Matrix4.translationValues(scrollController.offset, 0.0, 0.0),
+                      transform: Matrix4.translationValues(scrollController!.offset, 0.0, 0.0),
                       child: HomeDrawer(
                         screenIndex: widget.screenIndex == null ? DrawerIndex.HOME : widget.screenIndex,
                         iconAnimationController: iconAnimationController,
                         callBackIndex: (DrawerIndex indexType) {
                           onDrawerClick();
                           try {
-                            widget.onDrawerCall(indexType);
+                            widget.onDrawerCall!(indexType);
                           } catch (e) {}
                         },
                       ),
@@ -155,8 +155,8 @@ class _DrawerUserControllerState extends State<DrawerUserController> with Ticker
                                 child: widget.menuView != null
                                     ? widget.menuView
                                     : AnimatedIcon(
-                                        icon: widget.animatedIconData != null ? widget.animatedIconData : AnimatedIcons.arrow_menu,
-                                        progress: iconAnimationController),
+                                        icon: widget.animatedIconData ?? AnimatedIcons.arrow_menu,
+                                        progress: iconAnimationController!),
                               ),
                               onTap: () {
                                 FocusScope.of(context).requestFocus(FocusNode());
@@ -179,14 +179,14 @@ class _DrawerUserControllerState extends State<DrawerUserController> with Ticker
 
   void onDrawerClick() {
     //if scrollcontroller.offset != 0.0 then we set to closed the drawer(with animation to offset zero position) if is not 1 then open the drawer
-    if (scrollController.offset != 0.0) {
-      scrollController.animateTo(
+    if (scrollController!.offset != 0.0) {
+      scrollController?.animateTo(
         0.0,
         duration: const Duration(milliseconds: 400),
         curve: Curves.fastOutSlowIn,
       );
     } else {
-      scrollController.animateTo(
+      scrollController?.animateTo(
         widget.drawerWidth,
         duration: const Duration(milliseconds: 400),
         curve: Curves.fastOutSlowIn,
