@@ -12,7 +12,7 @@ class FitnessAppHomeScreen extends StatefulWidget {
 
 class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
     with TickerProviderStateMixin {
-  AnimationController animationController;
+  AnimationController? animationController;
 
   List<TabIconData> tabIconsList = TabIconData.tabIconsList;
 
@@ -35,7 +35,7 @@ class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
 
   @override
   void dispose() {
-    animationController.dispose();
+    animationController?.dispose();
     super.dispose();
   }
 
@@ -80,7 +80,7 @@ class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
           addClick: () {},
           changeIndex: (int index) {
             if (index == 0 || index == 2) {
-              animationController.reverse().then<dynamic>((data) {
+              animationController?.reverse().then<dynamic>((data) {
                 if (!mounted) {
                   return;
                 }
@@ -90,7 +90,7 @@ class _FitnessAppHomeScreenState extends State<FitnessAppHomeScreen>
                 });
               });
             } else if (index == 1 || index == 3) {
-              animationController.reverse().then<dynamic>((data) {
+              animationController?.reverse().then<dynamic>((data) {
                 if (!mounted) {
                   return;
                 }
