@@ -4,18 +4,18 @@ import '../fintness_app_theme.dart';
 
 class AreaListView extends StatefulWidget {
   const AreaListView(
-      {Key key, this.mainScreenAnimationController, this.mainScreenAnimation})
+      {Key? key, this.mainScreenAnimationController, this.mainScreenAnimation})
       : super(key: key);
 
-  final AnimationController mainScreenAnimationController;
-  final Animation<dynamic> mainScreenAnimation;
+  final AnimationController? mainScreenAnimationController;
+  final Animation<double>? mainScreenAnimation;
   @override
   _AreaListViewState createState() => _AreaListViewState();
 }
 
 class _AreaListViewState extends State<AreaListView>
     with TickerProviderStateMixin {
-  AnimationController animationController;
+  AnimationController? animationController;
   List<String> areaListData = <String>[
     'assets/fitness_app/area1.png',
     'assets/fitness_app/area2.png',
@@ -32,20 +32,20 @@ class _AreaListViewState extends State<AreaListView>
 
   @override
   void dispose() {
-    animationController.dispose();
+    animationController?.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: widget.mainScreenAnimationController,
-      builder: (BuildContext context, Widget child) {
+      animation: widget.mainScreenAnimationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: widget.mainScreenAnimation,
+          opacity: widget.mainScreenAnimation!,
           child: Transform(
             transform: Matrix4.translationValues(
-                0.0, 30 * (1.0 - widget.mainScreenAnimation.value), 0.0),
+                0.0, 30 * (1.0 - widget.mainScreenAnimation!.value), 0.0),
             child: AspectRatio(
               aspectRatio: 1.0,
               child: Padding(
@@ -62,16 +62,16 @@ class _AreaListViewState extends State<AreaListView>
                       final Animation<double> animation =
                           Tween<double>(begin: 0.0, end: 1.0).animate(
                         CurvedAnimation(
-                          parent: animationController,
+                          parent: animationController!,
                           curve: Interval((1 / count) * index, 1.0,
                               curve: Curves.fastOutSlowIn),
                         ),
                       );
-                      animationController.forward();
+                      animationController?.forward();
                       return AreaView(
                         imagepath: areaListData[index],
                         animation: animation,
-                        animationController: animationController,
+                        animationController: animationController!,
                       );
                     },
                   ),
@@ -93,26 +93,26 @@ class _AreaListViewState extends State<AreaListView>
 
 class AreaView extends StatelessWidget {
   const AreaView({
-    Key key,
+    Key? key,
     this.imagepath,
     this.animationController,
     this.animation,
   }) : super(key: key);
 
-  final String imagepath;
-  final AnimationController animationController;
-  final Animation<dynamic> animation;
+  final String? imagepath;
+  final AnimationController? animationController;
+  final Animation<double>? animation;
 
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: animationController,
-      builder: (BuildContext context, Widget child) {
+      animation: animationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: animation,
+          opacity: animation!,
           child: Transform(
             transform: Matrix4.translationValues(
-                0.0, 50 * (1.0 - animation.value), 0.0),
+                0.0, 50 * (1.0 - animation!.value), 0.0),
             child: Container(
               decoration: BoxDecoration(
                 color: FitnessAppTheme.white,
@@ -142,7 +142,7 @@ class AreaView extends StatelessWidget {
                       Padding(
                         padding:
                             const EdgeInsets.only(top: 16, left: 16, right: 16),
-                        child: Image.asset(imagepath),
+                        child: Image.asset(imagepath!),
                       ),
                     ],
                   ),
