@@ -4,28 +4,28 @@ import 'package:flutter/material.dart';
 import '../fintness_app_theme.dart';
 
 class GlassView extends StatelessWidget {
-  final AnimationController animationController;
-  final Animation animation;
+  final AnimationController? animationController;
+  final Animation<double>? animation;
 
-  const GlassView({Key key, this.animationController, this.animation})
+  const GlassView({Key? key, this.animationController, this.animation})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: animationController,
-      builder: (BuildContext context, Widget child) {
+      animation: animationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: animation,
+          opacity: animation!,
           child: new Transform(
             transform: new Matrix4.translationValues(
-                0.0, 30 * (1.0 - animation.value), 0.0),
+                0.0, 30 * (1.0 - animation!.value), 0.0),
             child: Column(
               children: <Widget>[
                 Padding(
                   padding: const EdgeInsets.only(
                       left: 24, right: 24, top: 0, bottom: 24),
                   child: Stack(
-                    overflow: Overflow.visible,
+                    clipBehavior: Clip.none,
                     children: <Widget>[
                       Padding(
                         padding: const EdgeInsets.only(top: 16),
