@@ -4,11 +4,11 @@ import 'package:flutter/material.dart';
 class TitleView extends StatelessWidget {
   final String titleTxt;
   final String subTxt;
-  final AnimationController animationController;
-  final Animation animation;
+  final AnimationController? animationController;
+  final Animation<double>? animation;
 
   const TitleView(
-      {Key key,
+      {Key? key,
       this.titleTxt: "",
       this.subTxt: "",
       this.animationController,
@@ -18,13 +18,13 @@ class TitleView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: animationController,
-      builder: (BuildContext context, Widget child) {
+      animation: animationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: animation,
+          opacity: animation!,
           child: new Transform(
             transform: new Matrix4.translationValues(
-                0.0, 30 * (1.0 - animation.value), 0.0),
+                0.0, 30 * (1.0 - animation!.value), 0.0),
             child: Container(
               child: Padding(
                 padding: const EdgeInsets.only(left: 24, right: 24),
