@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'model/homelist.dart';
 
 class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key key}) : super(key: key);
+  const MyHomePage({Key? key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -11,7 +11,7 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
   List<HomeList> homeList = HomeList.homeList;
-  AnimationController animationController;
+  AnimationController? animationController;
   bool multiple = true;
 
   @override
@@ -28,7 +28,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
   @override
   void dispose() {
-    animationController.dispose();
+    animationController?.dispose();
     super.dispose();
   }
 
@@ -69,12 +69,12 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                                 final Animation<double> animation =
                                     Tween<double>(begin: 0.0, end: 1.0).animate(
                                   CurvedAnimation(
-                                    parent: animationController,
+                                    parent: animationController!,
                                     curve: Interval((1 / count) * index, 1.0,
                                         curve: Curves.fastOutSlowIn),
                                   ),
                                 );
-                                animationController.forward();
+                                animationController?.forward();
                                 return HomeListView(
                                   animation: animation,
                                   animationController: animationController,
@@ -84,7 +84,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                                       context,
                                       MaterialPageRoute<dynamic>(
                                         builder: (BuildContext context) =>
-                                            homeList[index].navigateScreen,
+                                            homeList[index].navigateScreen!,
                                       ),
                                     );
                                   },
@@ -172,28 +172,28 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
 
 class HomeListView extends StatelessWidget {
   const HomeListView(
-      {Key key,
+      {Key? key,
       this.listData,
       this.callBack,
       this.animationController,
       this.animation})
       : super(key: key);
 
-  final HomeList listData;
-  final VoidCallback callBack;
-  final AnimationController animationController;
-  final Animation<dynamic> animation;
+  final HomeList? listData;
+  final VoidCallback? callBack;
+  final AnimationController? animationController;
+  final Animation<double>? animation;
 
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
-      animation: animationController,
-      builder: (BuildContext context, Widget child) {
+      animation: animationController!,
+      builder: (BuildContext context, Widget? child) {
         return FadeTransition(
-          opacity: animation,
+          opacity: animation!,
           child: Transform(
             transform: Matrix4.translationValues(
-                0.0, 50 * (1.0 - animation.value), 0.0),
+                0.0, 50 * (1.0 - animation!.value), 0.0),
             child: AspectRatio(
               aspectRatio: 1.5,
               child: ClipRRect(
@@ -202,7 +202,7 @@ class HomeListView extends StatelessWidget {
                   alignment: AlignmentDirectional.center,
                   children: <Widget>[
                     Image.asset(
-                      listData.imagePath,
+                      listData!.imagePath,
                       fit: BoxFit.cover,
                     ),
                     Material(
@@ -211,9 +211,7 @@ class HomeListView extends StatelessWidget {
                         splashColor: Colors.grey.withOpacity(0.2),
                         borderRadius:
                             const BorderRadius.all(Radius.circular(4.0)),
-                        onTap: () {
-                          callBack();
-                        },
+                        onTap: callBack,
                       ),
                     ),
                   ],
