@@ -5,19 +5,19 @@ class HotelAppTheme {
   static TextTheme _buildTextTheme(TextTheme base) {
     const String fontName = 'WorkSans';
     return base.copyWith(
-      headline1: base.headline1.copyWith(fontFamily: fontName),
-      headline2: base.headline2.copyWith(fontFamily: fontName),
-      headline3: base.headline3.copyWith(fontFamily: fontName),
-      headline4: base.headline4.copyWith(fontFamily: fontName),
-      headline5: base.headline5.copyWith(fontFamily: fontName),
-      headline6: base.headline6.copyWith(fontFamily: fontName),
-      button: base.button.copyWith(fontFamily: fontName),
-      caption: base.caption.copyWith(fontFamily: fontName),
-      bodyText1: base.bodyText1.copyWith(fontFamily: fontName),
-      bodyText2: base.bodyText2.copyWith(fontFamily: fontName),
-      subtitle1: base.subtitle1.copyWith(fontFamily: fontName),
-      subtitle2: base.subtitle2.copyWith(fontFamily: fontName),
-      overline: base.overline.copyWith(fontFamily: fontName),
+      headline1: base.headline1?.copyWith(fontFamily: fontName),
+      headline2: base.headline2?.copyWith(fontFamily: fontName),
+      headline3: base.headline3?.copyWith(fontFamily: fontName),
+      headline4: base.headline4?.copyWith(fontFamily: fontName),
+      headline5: base.headline5?.copyWith(fontFamily: fontName),
+      headline6: base.headline6?.copyWith(fontFamily: fontName),
+      button: base.button?.copyWith(fontFamily: fontName),
+      caption: base.caption?.copyWith(fontFamily: fontName),
+      bodyText1: base.bodyText1?.copyWith(fontFamily: fontName),
+      bodyText2: base.bodyText2?.copyWith(fontFamily: fontName),
+      subtitle1: base.subtitle1?.copyWith(fontFamily: fontName),
+      subtitle2: base.subtitle2?.copyWith(fontFamily: fontName),
+      overline: base.overline?.copyWith(fontFamily: fontName),
     );
   }
 
